refactor: move the update reference marker useEffect to the provider
